travel_bot/requirements.txt
python-telegram-bot==20.7
google-search-results
openai
python-dotenv
rich
